nvcfund-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
autocommand==2.2.2
babel==2.17.0
backports.tarfile==1.2.0
base58==2.1.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==4.2.1
bitarray==3.4.2
bleach==6.2.0
blinker==1.9.0
boolean.py==5.0
boto3>=1.35.0
botocore>=1.35.0
hvac==2.3.0
bots-ediint==3.3.1
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.3
cachelib==0.13.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cheroot==10.0.1
CherryPy==18.10.0
ckzg==2.1.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
courlan==1.3.2
crypto==1.4.1
cryptography==45.0.4
cssselect2==0.8.0
cyclonedx-python-lib==9.1.0
cytoolz==1.0.1
dateparser==1.2.2
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.2.3
django-user-management==18.0.0
djangorestframework==3.10.3
dnspython==2.7.0
docopt==0.6.2
docx==0.2.4
ecdsa==0.19.1
ECPy==1.2.5
email_validator==2.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
execnet==2.1.1
filelock==3.18.0
Flask==3.1.0
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-Dance==7.1.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.5
fpdf==1.7.2
frozenlist==1.7.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
hexbytes==1.3.1
htmldate==1.9.3
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib_metadata==8.7.0
incuna-mail==4.1.1
incuna-pigeon==0.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
jaraco.collections==5.2.1
jaraco.context==6.0.1
jaraco.functools==4.2.1
jaraco.text==4.0.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jusText==3.0.2
kombu==5.5.4
license-expression==30.4.3
limits==5.4.0
lxml==6.0.0
lxml_html_clean==0.4.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.5.0
Naked==0.1.32
nulltype==2.3.1
num2words==0.5.14
numpy==2.3.1
oauthlib==3.3.1
openai==1.93.0
ordered-set==4.1.0
packageurl-python==0.17.1
packaging==25.0
pandas==2.3.0
paramiko==3.5.1
parsimonious==0.10.0
paypalrestsdk==1.13.3
pdfkit==1.0.0
pillow==11.2.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip-review==1.3.0
pip-tools==7.4.1
pip_audit==2.9.0
plaid-python==34.0.0
platformdirs==4.3.8
playwright==1.53.0
pluggy==1.6.0
portend==3.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
py-ecc==8.0.0
py-serializable==2.0.0
py-solc-x==2.0.4
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.2
pyinotify==0.9.6
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyotp==2.9.0
pyparsing==3.2.3
pyphen==0.17.2
pyppeteer==2.0.0
pyproject_hooks==1.2.0
pysftp==0.2.9
pytest==8.4.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-http-client==3.3.7
python-socketio==5.13.0
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
regex==2024.11.6
reportlab==4.4.2
repoze.lru==0.7
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rlp==4.1.0
Routes==2.5.1
s3transfer==0.13.0
sendgrid==6.12.4
shellescape==3.8.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
stripe==12.2.0
tempora==5.8.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tld==0.13.1
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
trafilatura==2.0.0
twilio==9.6.3
types-Deprecated==1.2.15.20250304
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
URLObject==2.4.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
web3==7.12.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xrpl-py==2.4.0
yarl==1.20.1
zc.lockfile==3.0.post1
zipp==3.23.0
zopfli==0.2.3.post1
